requirements.txt
python-dotenv==1.0.1
termcolor==2.5.0
